Remove sdk-ext
- move sdk.Utils.Platform class to core Dicoogle,
  rename imports accordingly
- remove legacy constructs
  which have nothing to do with the current task system
   - remove `PluginController.addTask`
   - remove references to TaskQueue and TaskRequest
- remove unused imports in PluginController
