+++ dicoogle/pom.xml
@@ -375,12 +375,6 @@
         </dependency>
 
         <dependency>
-            <groupId>pt.ua.ieeta</groupId>
-            <artifactId>dicoogle-sdk-ext</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
             <groupId>org.eclipse.jetty</groupId>
             <artifactId>jetty-server</artifactId>
             <version>${jetty.version}</version>
