+++ dicoogle/src/main/java/pt/ua/dicoogle/core/QueryHistorySupport.java
@@ -28,7 +28,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.Observable;
 
-import pt.ua.dicoogle.sdk.Utils.Platform;
+import pt.ua.dicoogle.utils.Platform;
 
 /**
  *
