+++ dicoogle/src/main/java/pt/ua/dicoogle/core/settings/ServerSettingsManager.java
@@ -25,7 +25,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xml.sax.SAXException;
 import pt.ua.dicoogle.core.XMLSupport;
-import pt.ua.dicoogle.sdk.Utils.Platform;
+import pt.ua.dicoogle.utils.Platform;
 import pt.ua.dicoogle.sdk.settings.server.ServerSettings;
 
 import java.io.*;
