+++ dicoogle/src/main/java/pt/ua/dicoogle/server/web/servlets/management/LoggerServlet.java
@@ -36,7 +36,7 @@ import org.apache.logging.log4j.core.appender.RollingFileAppender;
 import org.apache.logging.log4j.core.appender.RollingRandomAccessFileAppender;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import pt.ua.dicoogle.sdk.Utils.Platform;
+import pt.ua.dicoogle.utils.Platform;
 
 /**
  * @author Frederico Silva <fredericosilva@ua.pt>
