--- sdk-ext/.gitignore
@@ -1,2 +1,0 @@
-/target
-/dependency-reduced-pom.xml
