--- sdk-ext/src/main/java/pt/ua/dicoogle/sdk/index/handlers/handler.properties
@@ -1,1 +1,0 @@
-dcm = pt.ua.dicoogle.core.index.LuceneSupport.DICOM.DicomDocumentNG
